[seller] refactor(144): 리뷰 피드백 반영하여 FSD 구조 정리
- entity/auth, features/auth 각 슬라이스에 index.ts 추가 (Public API)
- useIssueTokenMutation, useLogoutMutation을 entity -> features로 이동
- features/auth/ui 제거, auth-page.tsx에 레이아웃 JSX 인라인화
- social-callback-page: useEffect deps 의도 설명 주석 추가
- apps/admin/.env.example 원복 (충돌 해결 중 누락분)

features/auth/login 평탄화 및 파일 접미사 적용
- login-image.tsx → login-image.ui.tsx (named export 전환)
- social-login-buttons.tsx -> social-login-buttons.ui.tsx
- login-footer.tsx -> login-footer.ui.tsx
- constant/message.ts -> login.constant.ts
- 빈 login-hooks.ts 삭제
