--- apps/seller/src/features/auth/login/login-hooks.ts
@@ -1,1 +1,0 @@
-export {}
