+++ apps/seller/src/features/auth/login/login.mutation.ts
@@ -1,10 +1,7 @@
 import { useMutation } from '@tanstack/react-query'
 
+import { authKeys, issueToken, useAuthStore } from '@/entity/auth'
 import { getExpFromToken, setCookie } from '@/shared/utils/cookieUtils'
-
-import { issueToken, logout } from './auth-api'
-import { useAuthStore } from './auth-store'
-import { authKeys } from './key'
 
 export const useIssueTokenMutation = () => {
   const setAuth = useAuthStore((state) => state.setAuth)
@@ -24,19 +21,6 @@ export const useIssueTokenMutation = () => {
       if (data.result) {
         setAuth(data.result.sellerId, data.result.status)
       }
-    },
-  })
-}
-
-export const useLogoutMutation = () => {
-  const logoutStore = useAuthStore((state) => state.logout)
-
-  return useMutation({
-    mutationKey: [...authKeys.all, 'logout'],
-    mutationFn: () => logout(),
-    onSuccess: () => {
-      logoutStore()
-      window.location.href = '/auth'
     },
   })
 }
