+++ apps/seller/src/features/auth/logout/index.ts
@@ -1,0 +1,1 @@
+export { useLogoutMutation } from './logout.mutation'
