--- apps/seller/src/features/auth/ui/auth-card.tsx
@@ -1,21 +1,0 @@
-import { ReactNode } from 'react'
-
-import { clsx } from 'clsx'
-
-interface CardProps {
-  children: ReactNode
-  className?: string
-}
-
-export const AuthCard = ({ children, className = '' }: CardProps) => {
-  return (
-    <div
-      className={clsx(
-        'flex flex-1 items-center gap-4 self-stretch overflow-hidden rounded-20 border border-gray-200 bg-white p-2.5',
-        className,
-      )}
-    >
-      {children}
-    </div>
-  )
-}
