--- apps/seller/src/features/auth/ui/index.ts
@@ -1,3 +1,0 @@
-export { AuthCard } from './auth-card'
-export { AuthContentWrapper } from './auth-content-wrapper'
-export { AuthPageContainer } from './auth-page-container'
